Use 64-bit uint instead of 32-bit for sequence number in packets to simplify logic.
I was going to use 32-bit unsigned values and implement logic for wrap around,
but I'm going to keep it simple instead. Now we assume that wrap around will never be
necessary, as it is safe to assume that no one sender will send 2^64 bytes from one
continually running process.

+++ header/header_test.go
@@ -7,7 +7,7 @@ import (
 )
 
 func TestHeaderTypes(t *testing.T) {
-	const seq = uint32(23)
+	const seq = uint64(23)
 	message := MakeMessageHeader(seq)
 	request := MakeRequestHeader()
 	response := MakeResponseHeader()
@@ -87,7 +87,7 @@ func TestHeaderTypes(t *testing.T) {
 
 
 func TestMakeMessageHeader(t *testing.T) {
-	const seq = uint32(23)
+	const seq = uint64(23)
 	h := MakeMessageHeader(seq)
 
 	if !h.Valid() {
